pyproject.toml
[tool.poetry]
name = "lewidi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "lewidi", from = "src" }]

[tool.poe]
envfile = ".env"

[tool.poe.tasks]

[tool.poe.tasks.format]
help = "Format using the pre-commit hooks"
cmd = "pre-commit run --all-files"

[tool.poe.tasks.typecheck]
help = "Check types with mypy"
cmd = "mypy ."

[tool.poe.tasks.lint]
help = "Lint with flake8"
cmd = "flake8 ."

[tool.poetry.dependencies]
python = "~3.10"
pydantic = "^1.10.4"
torch = "^1.13.1"
orjson = "^3.8.5"
scikit-learn = "^1.2.1"
transformers = "^4.26.0"
rich = "^13.2.0"
pytorch-lightning = "^1.9.0"
torchdata = "^0.5.1"
arabert = "^1.0.1"
allennlp-light = "^1.0.0"
jaxtyping = "^0.2.11"
ai2-tango = "^1.2.0"
typeguard = "^2.13.3"
hydra-zen = "^0.9.1"
loguru = "^0.6.0"
wandb = "^0.13.10"

[tool.poetry.group.lint.dependencies]
wemake-python-styleguide = "^0.17.0"
mypy = "^0.991"
pytest = "^7.2.1"
pre-commit = "^3.0.0"
poethepoet = "^0.18.1"
ruff = "^0.0.252"
black = "^23.1.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-cases = "^3.6.13"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
combine_as_imports = true
lines_after_imports = 2
line_length = 99
known_first_party = ["lewidi"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
